fixed ClassCastException by replacing (Integer) with (Long)
git-svn-id: https://server.mycore.de/svn/mycore/trunk@11623 7089d063-bc14-496d-ac90-53de1872eb6c

+++ sources/org/mycore/backend/hibernate/MCRHIBAccessStore.java
@@ -152,7 +152,7 @@ public class MCRHIBAccessStore extends MCRAccessStore {
         }
         
         try {
-        	int count = ( (Integer) session.createQuery(query.toString()).iterate().next() ).intValue();
+        	int count = ( (Long) session.createQuery(query.toString()).iterate().next() ).intValue();
         	if (count > 0){
         		return true;
         	}
